feat: add bottom tab navigation
In the above code, we have created a tab navigator using the  createBottomTabNavigator  function. We have created three screens, Home, Settings, and Notifications. We have used the  tabBarIcon  option to display the icon for each tab. We have used the  tabBarActiveTintColor  and  tabBarInactiveTintColor  options to set the color of the active and inactive tabs.